requirements.txt
aiogram
aiohttp
motor
python-dotenv
Pillow
google-generativeai
flask